Refactor Biblioteca and Juego classes to use long for ID fields; update repository methods and add HTML templates for login and main page

+++ src/main/java/com/es/unex/cum/mdai/Steem/Modelo/Juego.java
@@ -33,7 +33,7 @@ public class Juego {
         return idJuego;
     }
 
-    public void setIdJuego(int idJuego) {
+    public void setIdJuego(long idJuego) {
         this.idJuego = idJuego;
     }
